Separate build file
